requirements.txt
absl-py==2.2.0
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Flask==2.3.3
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==2.14.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
ml-dtypes==0.2.0
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.23.5
oauthlib==3.2.0
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==21.3
pexpect==4.8.0
pillow==10.2.0
protobuf==4.25.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.13.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.8.2
ssh-import-id==5.11
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
Twisted==22.1.0
typeguard==2.13.3
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
validators==0.34.0
vit-keras==0.1.2
wadllib==1.3.6
waitress==2.1.2
Werkzeug==3.1.3
wrapt==1.14.1
zipp==1.0.0
zope.interface==5.4.0
